move the cdk code to the root directory
