--- cdk/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "cdk",
-  "version": "0.1.0",
-  "bin": {
-    "cdk": "bin/cdk.js"
-  },
-  "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "test": "jest",
-    "cdk": "cdk"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.12",
-    "@types/node": "20.14.2",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.4",
-    "aws-cdk": "2.147.0",
-    "ts-node": "^10.9.2",
-    "typescript": "~5.4.5"
-  },
-  "dependencies": {
-    "aws-cdk-lib": "2.147.0",
-    "constructs": "^10.0.0",
-    "source-map-support": "^0.5.21"
-  }
-}
