+++ lib/base-infra-stack.ts
@@ -1,6 +1,9 @@
 import * as cdk from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import * as ec2 from "aws-cdk-lib/aws-ec2";
+import * as s3 from "aws-cdk-lib/aws-s3";
+import * as s3deploy from 'aws-cdk-lib/aws-s3-deployment';
+import path = require("path");
 
 
 export class BaseInfraStack extends cdk.Stack {
@@ -15,7 +18,7 @@ export class BaseInfraStack extends cdk.Stack {
 
     NOTE - the AWS profile that is used to deploy should have the same default region
     */
-    const regionPrefix = process.env.CDK_DEFAULT_REGION || 'eu-north-1';
+    const regionPrefix = process.env.CDK_DEFAULT_REGION || 'us-east-1';
     console.log(`CDK_DEFAULT_REGION: ${regionPrefix}`);
 
     // create VPC to deploy the infrastructure in
@@ -44,6 +47,14 @@ export class BaseInfraStack extends cdk.Stack {
         allowAllOutbound: true
     });
     this.redshiftServerlessSecGroup = redshiftServerlessSg;
+
+    // create bucket for synthetic dataset
+    const dataBucket = new s3.Bucket(this, `syntheticData`, {});
+    // use s3 bucket deploy to upload the local copy of the synthetic data file to s3
+    new s3deploy.BucketDeployment(this, 'dataBucketDeploy', {
+        sources: [s3deploy.Source.asset(path.join(__dirname, "../synthetic-dataset"))],
+        destinationBucket: dataBucket
+    });
 
   }
 }
